+++ changelog.md
@@ -13,6 +13,8 @@ Codion Change Log
 - FilterModel.visible() and filtered() renamed included() and excluded(), related renaming.
 - FilterModel.AbstractRefresher.supplier() renamed items().
 - FilterModel.Builder.include() renamed included(), related renaming.
+### is.codion.swing.common.model
+- ProgressWorker, internal interfaces sealed.
 ### is.codion.swing.common.ui
 - FilterTable.Builder.rowSelection() and columnSelection() added.
 - ControlBuilder.icon() parameter now nullable.